#7 Implement default shipping rate calculation that can be overriden with custom implementation if required
