+++ Composers/RegisterServicesComposer.cs
@@ -17,6 +17,7 @@ namespace UmbCheckout.Stripe.Composers
             builder.Services.AddTransient<IStripeSessionService, StripeSessionService>();
             builder.Services.AddTransient<IStripeShippingRateDatabaseService, StripeShippingRateDatabaseService>();
             builder.Services.AddTransient<IStripeShippingRateApiService, StripeShippingRateApiService>();
+            builder.Services.AddTransient<IStripeShippingOptionsService, StripeDefaultShippingOptionsService>();
             builder.Services.AddTransient<IStripeSettingsService, StripeSettingsService>();
 
 #if NET8_0_OR_GREATER
