+++ Services/StripeSessionService.cs
@@ -29,11 +29,11 @@ namespace UmbCheckout.Stripe.Services
         private readonly IConfigurationService _configurationService;
         private readonly ILogger<StripeSessionService> _logger;
         private readonly IHttpContextAccessor _httpContextAccessor;
-        private readonly IStripeShippingRateDatabaseService _stripeDatabaseService;
+        private readonly IStripeShippingOptionsService _stripeShippingOptionsService;
         private readonly StripeSettings _stripeSettings;
         private readonly IStripeSettingsService _stripeSettingsService;
 
-        public StripeSessionService(IPublishedSnapshotAccessor snapshotAccessor, ICoreScopeProvider coreScopeProvider, IEventAggregator eventAggregator, ILogger<StripeSessionService> logger, IConfigurationService configurationService, IHttpContextAccessor httpContextAccessor, IStripeShippingRateDatabaseService stripeDatabaseService, IOptionsMonitor<StripeSettings> stripeSettings, LicenseService licenseService, IStripeSettingsService stripeSettingsService)
+        public StripeSessionService(IPublishedSnapshotAccessor snapshotAccessor, ICoreScopeProvider coreScopeProvider, IEventAggregator eventAggregator, ILogger<StripeSessionService> logger, IConfigurationService configurationService, IHttpContextAccessor httpContextAccessor, IStripeShippingOptionsService stripeShippingOptionsService, IOptionsMonitor<StripeSettings> stripeSettings, LicenseService licenseService, IStripeSettingsService stripeSettingsService)
         {
             _snapshotAccessor = snapshotAccessor;
             _coreScopeProvider = coreScopeProvider;
@@ -41,7 +41,7 @@ namespace UmbCheckout.Stripe.Services
             _logger = logger;
             _configurationService = configurationService;
             _httpContextAccessor = httpContextAccessor;
-            _stripeDatabaseService = stripeDatabaseService;
+            _stripeShippingOptionsService = stripeShippingOptionsService;
             _stripeSettingsService = stripeSettingsService;
             _stripeSettings = stripeSettings.CurrentValue;
             licenseService.RunLicenseCheck();
@@ -390,16 +390,7 @@ namespace UmbCheckout.Stripe.Services
 
                     if (configuration is { EnableShipping: true })
                     {
-                        var shippingRates = await _stripeDatabaseService.GetShippingRates();
-                        var shippingRatesList = shippingRates.ToList();
-
-                        if (shippingRatesList.Any())
-                        {
-                            var shippingRateOptions = shippingRatesList.Select(shippingRate =>
-                                new SessionShippingOptionOptions { ShippingRate = shippingRate.Value }).ToList();
-
-                            options.ShippingOptions = shippingRateOptions;
-                        }
+                        options.ShippingOptions = await _stripeShippingOptionsService.GetShippingOptions(basket);
 
                         options.BillingAddressCollection = "required";
